requirements.txt
tensorflow
numpy
opencv-python
scikit-learn
fastapi
uvicorn
python-multipart
streamlit
pillow
requests